+++ src/hit.zig
@@ -1,7 +1,7 @@
 const vec3 = @import("vec3.zig");
 
 pub const Hit = struct {
-    t: f32,
+    t: f64,
     p: vec3.Vec3,
     n: vec3.Vec3,
 };
